LayoutTests/imported/w3c/web-platform-tests/scroll-animations/scroll-timelines/setting-timeline.tentative-expected.txt


PASS Setting a scroll timeline on a play-pending animation synchronizes currentTime of the animation with the scroll position.
PASS Setting a scroll timeline on a pause-pending animation fixes the currentTime of the animation based on the scroll position once resumed
PASS Setting a scroll timeline on a reversed play-pending animation synchronizes the currentTime of the animation with the scroll position.
FAIL Setting a scroll timeline on a running animation synchronizes the currentTime of the animation with the scroll position. assert_approx_equals: Timeline's currentTime aligns with the scroll position even when paused expected a number but got a "object"
PASS Setting a scroll timeline on a paused animation fixes the currentTime of the animation based on the scroll position when resumed
PASS Setting a scroll timeline on a reversed paused animation fixes the currentTime of the animation based on the scroll position when resumed
PASS Transitioning from a scroll timeline to a document timeline on a running animation preserves currentTime
PASS Transitioning from a scroll timeline to a document timeline on a pause-pending animation preserves currentTime
FAIL Transition from a scroll timeline to a document timeline on a reversed paused animation maintains correct currentTime assert_approx_equals: values do not match for "Animation's currentTime aligns with the scroll position" expected 90 +/- 0.125 but got -100
PASS Transitioning from a scroll timeline to a null timeline on a running animation preserves current progress.
FAIL Switching from a null timeline to a scroll timeline on an animation with a resolved start time preserved the play state promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'actual.unit')"
FAIL Switching from one scroll timeline to another updates currentTime promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'actual.unit')"
PASS Switching from a document timeline to a scroll timeline updates currentTime when unpaused via CSS.
PASS Switching from a document timeline to a scroll timeline and updating currentTime preserves the progress while paused.
FAIL Switching from a document timeline to a scroll timeline on an infinite duration animation. assert_equals: 'actual' unit type must be 'percent' for "undefined" expected (string) "percent" but got (undefined) undefined
FAIL Changing from a scroll-timeline to a view-timeline updates start time. assert_approx_equals: Timeline's currentTime aligns with the scroll position even when paused expected a number but got a "object"
